// src/server/db-types.ts
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

import type { ColumnType } from "kysely";

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export type Int8 = ColumnType<string, bigint | number | string, bigint | number | string>;

export type Json = JsonValue;

export type JsonArray = JsonValue[];

export type JsonObject = {
  [x: string]: JsonValue | undefined;
};

export type JsonPrimitive = boolean | number | string | null;

export type JsonValue = JsonArray | JsonObject | JsonPrimitive;

export type Timestamp = ColumnType<Date, Date | string, Date | string>;

export interface Assets {
  hash: string;
  id: Generated<Int8>;
  mtime: Timestamp;
  path: string;
  scanned_at: Generated<Timestamp>;
  size: Int8;
}

export interface Pgmigrations {
  id: Generated<number>;
  name: string;
  run_on: Timestamp;
}

export interface RemoteAssets {
  hash: string;
  id: Generated<Int8>;
  path: string;
  size: Int8;
  synced_at: Generated<Timestamp>;
}

export interface SyncRuns {
  added_count: Generated<number>;
  deleted_count: Generated<number>;
  finished_at: Timestamp | null;
  generated_macro: string | null;
  id: Generated<Int8>;
  modified_count: Generated<number>;
  outcome: Generated<string>;
  purged_urls: Generated<Json>;
  renamed_count: Generated<number>;
  started_at: Generated<Timestamp>;
  world_acknowledgements: Generated<Json>;
}

export interface DB {
  assets: Assets;
  pgmigrations: Pgmigrations;
  remote_assets: RemoteAssets;
  sync_runs: SyncRuns;
}